+++ bot/bot_funcs/__init__.py
@@ -1,5 +1,6 @@
 # Python package for bot command and functionality 
 from bot_funcs.message_listener import add_or_remove_signup_reactions
+from bot_funcs.message_listener import send_event_announcement
 from bot_funcs.dm_manager import send_dm_to_new_member
 from bot_funcs.commands.enlist_member import enlist_member
 from bot_funcs.commands.grant_role import grant_role